Clean up docs for contrib.forecast (#2326)
* Clean up docs for contrib.forecast

* Fix typos, add links
